feat: change to arena header
