+++ src/containers/arenaChat.tsx
@@ -7,8 +7,6 @@ import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { chatResult, chatReward, chatWithModel } from "@/lib/arena";
 import { ArenaStatus, ChoiceType } from "@/type";
-import { requestAddress } from "@/lib/wallet";
-import { WalletOutlined } from "@ant-design/icons";
 import ChoiceButton from "@/components/choiceButton";
 
 type ArenaChatProps = {
@@ -27,8 +25,6 @@ export default function ArenaChat({modelA, modelB}: ArenaChatProps) {
   const [resultA, setResultA] = useState("");
   const [resultB, setResultB] = useState("");
   const [notiApi, notiContextHolder] = notification.useNotification();
-  const [modalLoading, setModalLoading] = useState(false);
-  const [modalOpen, setModalOpen] = useState(false);
   
   const openNotification = (rewardData: any) => {
     console.log(rewardData);
@@ -64,22 +60,6 @@ export default function ArenaChat({modelA, modelB}: ArenaChatProps) {
     setStatus(ArenaStatus.READY);
     setResultA("");
     setResultB("");
-  }
-
-  const handleOk = () => {
-    setModalOpen(false);
-  };
-
-  const onClickConnectWalletBtn = async () => {
-    if(address !== "") {
-      return;
-    }
-    const getAddress = await requestAddress();
-    if(getAddress === undefined){
-      setModalOpen(true);
-      return;
-    }
-    setAddress(getAddress)
   }
 
   const onClickChoiceBtn = async (value: ChoiceType) => {
@@ -138,26 +118,6 @@ export default function ArenaChat({modelA, modelB}: ArenaChatProps) {
   return (
     <Space direction="vertical" size="middle" style={{ display: 'flex' }}>
       {notiContextHolder}
-      <Modal
-        open={modalOpen}
-        title="Ain Wallet not found"
-        onOk={handleOk}
-        onCancel={handleOk}
-        footer={[
-          <Button
-            key="link"
-            href="https://chromewebstore.google.com/detail/ain-wallet/hbdheoebpgogdkagfojahleegjfkhkpl?hl=ko"
-            type="primary"
-            loading={modalLoading}
-            onClick={handleOk}
-          >
-            Install AIN wallet extention
-          </Button>,
-        ]}
-      >
-        <p>You should install AIN wallet first.</p>
-      </Modal>
-      <Button onClick={onClickConnectWalletBtn} ><WalletOutlined />{address ? address.slice(0,8)+"..." : "connect wallet"} </Button>
       <Flex justify="space-between">
         <ChatBox modelName={modelAName} status={status} prompt={resultA} />
         <ChatBox modelName={modelBName} status={status} prompt={resultB} />
