+++ src/containers/recoilRootWrapper.tsx
@@ -1,20 +1,12 @@
 "use client";
 
+import ArenaHeader from "@/components/arenaHeader";
 import { Content, Header } from "antd/lib/layout/layout";
 import { RecoilRoot } from "recoil";
 
 type RecoilRootWrapperProps = {
   children: React.ReactNode,
 }
-
-const headerStyle: React.CSSProperties = {
-  textAlign: "center",
-  fontSize: "48px",
-  height: 64,
-  paddingInline: 48,
-  lineHeight: "64px",
-  backgroundColor: "#ffffff",
-};
 
 const contentStyle: React.CSSProperties = {
   textAlign: "center",
@@ -25,7 +17,7 @@ const contentStyle: React.CSSProperties = {
 export default function RecoilRootWrapper({ children }: RecoilRootWrapperProps) {
   return (
     <RecoilRoot>
-      <Header style={headerStyle}>AI Network Chatbot Arena</Header>
+      <ArenaHeader />
       <Content style={contentStyle}>
           {children}
       </Content>
